+++ include/boost/beast/websocket/stream.hpp
@@ -21,6 +21,7 @@
 #include <boost/beast/websocket/detail/mask.hpp>
 #include <boost/beast/websocket/detail/pausation.hpp>
 #include <boost/beast/websocket/detail/pmd_extension.hpp>
+#include <boost/beast/websocket/detail/stream_base.hpp>
 #include <boost/beast/websocket/detail/utf8_checker.hpp>
 #include <boost/beast/core/static_buffer.hpp>
 #include <boost/beast/core/string.hpp>
@@ -29,8 +30,6 @@
 #include <boost/beast/http/message.hpp>
 #include <boost/beast/http/string_body.hpp>
 #include <boost/beast/http/detail/type_traits.hpp>
-#include <boost/beast/zlib/deflate_stream.hpp>
-#include <boost/beast/zlib/inflate_stream.hpp>
 #include <boost/asio/async_result.hpp>
 #include <boost/asio/error.hpp>
 #include <algorithm>
@@ -108,6 +107,12 @@ class frame_test;
     For asynchronous operations, the type must support the
     @b AsyncStream concept.
 
+    @tparam deflateSupported A `bool` indicating whether or not the
+    stream will be capable of negotiating the permessage-deflate websocket
+    extension. Note that even if this is set to `true`, the permessage
+    deflate options (set by the caller at runtime) must still have the
+    feature enabled for a successful negotiation to occur.
+
     @note A stream object must not be moved or destroyed while there
     are pending asynchronous operations associated with it.
 
@@ -116,8 +121,13 @@ class frame_test;
         @b DynamicBuffer,
         @b SyncStream
 */
-template<class NextLayer>
+template<
+    class NextLayer,
+    bool deflateSupported>
 class stream
+#ifndef BOOST_BEAST_DOXYGEN
+    : private detail::stream_base<deflateSupported>
+#endif
 {
     friend class close_test;
     friend class frame_test;
@@ -126,15 +136,13 @@ class stream
     friend class read2_test;
     friend class stream_test;
     friend class write_test;
-
+    
     /*  The read buffer has to be at least as large
         as the largest possible control frame including
         the frame header.
     */
     static std::size_t constexpr max_control_frame_size = 2 + 8 + 4 + 125;
     static std::size_t constexpr tcp_frame_size = 1536;
-
-    struct op {};
 
     using control_cb_type =
         std::function<void(frame_type, string_view)>;
@@ -152,16 +160,6 @@ class stream
         bool operator!=(token const& t) { return id_ != t.id_; }
         token unique() { token t{id_++}; if(id_ == 0) ++id_; return t; }
         void reset() { id_ = 0; }
-    };
-
-    // State information for the permessage-deflate extension
-    struct pmd_t
-    {
-        // `true` if current read message is compressed
-        bool rd_set = false;
-
-        zlib::deflate_stream zo;
-        zlib::inflate_stream zi;
     };
 
     enum class status
@@ -231,14 +229,14 @@ class stream
     detail::pausation       paused_wr_;     // paused write op
     detail::pausation       paused_ping_;   // paused ping op
     detail::pausation       paused_close_;  // paused close op
-    detail::pausation       paused_r_rd_;   // paused read op (read)
-    detail::pausation       paused_r_close_;// paused close op (read)
-
-    std::unique_ptr<pmd_t>  pmd_;           // pmd settings or nullptr
-    permessage_deflate      pmd_opts_;      // local pmd options
-    detail::pmd_offer       pmd_config_;    // offer (client) or negotiation (server)
+    detail::pausation       paused_r_rd_;   // paused read op (async read)
+    detail::pausation       paused_r_close_;// paused close op (async read)
 
 public:
+    /// Indicates if the permessage-deflate extension is supported
+    using is_deflate_supported =
+        std::integral_constant<bool, deflateSupported>;
+
     /// The type of the next layer.
     using next_layer_type =
         typename std::remove_reference<NextLayer>::type;
@@ -444,7 +442,11 @@ public:
     */
     std::size_t
     read_size_hint(
-        std::size_t initial_size = +tcp_frame_size) const;
+        std::size_t initial_size = +tcp_frame_size) const
+    {
+        return read_size_hint(initial_size,
+            is_deflate_supported{});
+    }
 
     /** Returns a suggested maximum buffer size for the next call to read.
 
@@ -475,15 +477,22 @@ public:
     //
     //--------------------------------------------------------------------------
 
-    /// Set the permessage-deflate extension options
+    /** Set the permessage-deflate extension options
+
+        @throws invalid_argument if `deflateSupported == false`, and either
+        `client_enable` or `server_enable` is `true`.
+    */
     void
-    set_option(permessage_deflate const& o);
+    set_option(permessage_deflate const& o)
+    {
+        set_option(o, is_deflate_supported{});
+    }
 
     /// Get the permessage-deflate extension options
     void
     get_option(permessage_deflate& o)
     {
-        o = pmd_opts_;
+        get_option(o, is_deflate_supported{});
     }
 
     /** Set the automatic fragmentation option.
@@ -2353,8 +2362,8 @@ public:
         next layer's `async_write_some` functions, and is known as a
         <em>composed operation</em>. The program must ensure that the
         stream performs no other write operations (such as @ref async_ping,
-        @ref stream::async_write, @ref stream::async_write_some, or
-        @ref stream::async_close) until this operation completes.
+        @ref async_write, @ref async_write_some, or @ref async_close)
+        until this operation completes.
 
         If the close reason specifies a close code other than
         @ref beast::websocket::close_code::none, the close frame is
@@ -3172,8 +3181,8 @@ public:
         to the next layer's `async_write_some` functions, and is known
         as a <em>composed operation</em>. The program must ensure that
         the stream performs no other write operations (such as
-        stream::async_write, stream::async_write_some, or
-        stream::async_close).
+        @ref async_write, @ref async_write_some, or
+        @ref async_close).
 
         The current setting of the @ref binary option controls
         whether the message opcode is set to text or binary. If the
@@ -3301,8 +3310,8 @@ public:
         as a <em>composed operation</em>. The actual payload sent
         may be transformed as per the WebSocket protocol settings. The
         program must ensure that the stream performs no other write
-        operations (such as stream::async_write, stream::async_write_some,
-        or stream::async_close).
+        operations (such as @ref async_write, @ref async_write_some,
+        or @ref async_close).
 
         If this is the beginning of a new message, the message opcode
         will be set to text or binary as per the current setting of
@@ -3351,10 +3360,65 @@ private:
     static void default_decorate_req(request_type&) {}
     static void default_decorate_res(response_type&) {}
 
+    void
+    set_option(permessage_deflate const& o, std::true_type);
+
+    void
+    set_option(permessage_deflate const&, std::false_type);
+
+    void
+    get_option(permessage_deflate& o, std::true_type)
+    {
+        o = this->pmd_opts_;
+    }
+
+    void
+    get_option(permessage_deflate& o, std::false_type)
+    {
+        o = {};
+        o.client_enable = false;
+        o.server_enable = false;
+    }
+
     void open(role_type role);
+
+    void open_pmd(std::true_type);
+
+    void open_pmd(std::false_type)
+    {
+    }
+
     void close();
+
+    void close_pmd(std::true_type)
+    {
+        this->pmd_.reset();
+    }
+
+    void close_pmd(std::false_type)
+    {
+    }
+    
     void reset();
-    void begin_msg();
+
+    void begin_msg()
+    {
+        begin_msg(is_deflate_supported{});
+    }
+
+    void begin_msg(std::true_type);
+
+    void begin_msg(std::false_type);
+
+    std::size_t
+    read_size_hint(
+        std::size_t initial_size,
+        std::true_type) const;
+
+    std::size_t
+    read_size_hint(
+        std::size_t initial_size,
+        std::false_type) const;
 
     bool
     check_open(error_code& ec)
@@ -3394,6 +3458,10 @@ private:
     write_ping(DynamicBuffer& b,
         detail::opcode op, ping_data const& data);
 
+    //
+    // upgrade
+    //
+
     template<class Decorator>
     request_type
     build_request(detail::sec_ws_key_type& key,
@@ -3401,28 +3469,94 @@ private:
             string_view target,
                 Decorator const& decorator);
 
-    template<class Body,
-        class Allocator, class Decorator>
-    response_type
-    build_response(http::request<Body,
-        http::basic_fields<Allocator>> const& req,
-            Decorator const& decorator);
+    void
+    build_request_pmd(request_type& req, std::true_type);
 
     void
-    on_response(response_type const& resp,
-        detail::sec_ws_key_type const& key, error_code& ec);
+    build_request_pmd(request_type&, std::false_type)
+    {
+    }
+
+    template<
+        class Body, class Allocator, class Decorator>
+    response_type
+    build_response(
+        http::request<Body,
+            http::basic_fields<Allocator>> const& req,
+        Decorator const& decorator);
+
+    template<class Body, class Allocator>
+    void
+    build_response_pmd(
+        response_type& res,
+        http::request<Body,
+            http::basic_fields<Allocator>> const& req,
+        std::true_type);
+
+    template<class Body, class Allocator>
+    void
+    build_response_pmd(
+        response_type&,
+        http::request<Body,
+            http::basic_fields<Allocator>> const&,
+        std::false_type)
+    {
+    }
+
+    void
+    on_response(
+        response_type const& res,
+        detail::sec_ws_key_type const& key,
+        error_code& ec);
+
+    void
+    on_response_pmd(
+        response_type const& res,
+        std::true_type);
+
+    void
+    on_response_pmd(
+        response_type const&,
+        std::false_type)
+    {
+    }
+
+    //
+    // accept / handshake
+    //
+
+    template<class Allocator>
+    void
+    do_pmd_config(
+        http::basic_fields<Allocator> const& h,
+        std::true_type)
+    {
+        pmd_read(this->pmd_config_, h);
+    }
+
+    template<class Allocator>
+    void
+    do_pmd_config(
+        http::basic_fields<Allocator> const&,
+        std::false_type)
+    {
+    }
 
     template<class Decorator>
     void
-    do_accept(Decorator const& decorator,
+    do_accept(
+        Decorator const& decorator,
         error_code& ec);
 
-    template<class Body, class Allocator,
+    template<
+        class Body, class Allocator,
         class Decorator>
     void
-    do_accept(http::request<Body,
-        http::basic_fields<Allocator>> const& req,
-            Decorator const& decorator, error_code& ec);
+    do_accept(
+        http::request<Body,
+            http::basic_fields<Allocator>> const& req,
+        Decorator const& decorator,
+        error_code& ec);
 
     template<class RequestDecorator>
     void
@@ -3430,6 +3564,10 @@ private:
         string_view host, string_view target,
             RequestDecorator const& decorator,
                 error_code& ec);
+
+    //
+    // fail
+    //
 
     void
     do_fail(