+++ include/boost/beast/websocket/stream_fwd.hpp
@@ -17,7 +17,8 @@ namespace beast {
 namespace websocket {
 
 template<
-    class NextLayer>
+    class NextLayer,
+    bool deflateSupported = true>
 class stream;
 
 } // websocket