+++ test/beast/websocket/test.hpp
@@ -35,6 +35,10 @@ class websocket_test_suite
     , public test::enable_yield_to
 {
 public:
+    template<bool deflateSupported>
+    using ws_type_t =
+        websocket::stream<test::stream&, deflateSupported>;
+
     using ws_type =
         websocket::stream<test::stream&>;
 
@@ -303,7 +307,7 @@ public:
             , limit);
     }
 
-    template<class Test>
+    template<bool deflateSupported = true, class Test>
     void
     doTest(
         permessage_deflate const& pmd,
@@ -320,7 +324,7 @@ public:
             {
                 test::fail_counter fc{n};
                 test::stream ts{ioc_, fc};
-                ws_type ws{ts};
+                ws_type_t<deflateSupported> ws{ts};
                 ws.set_option(pmd);
 
                 echo_server es{log, i==1 ?
@@ -481,146 +485,171 @@ public:
 
     struct SyncClient
     {
-        template<class NextLayer>
+        template<class NextLayer, bool deflateSupported>
         void
-        accept(stream<NextLayer>& ws) const
+        accept(
+            stream<NextLayer, deflateSupported>& ws) const
         {
             ws.accept();
         }
 
-        template<class NextLayer, class Buffers>
+        template<
+            class NextLayer, bool deflateSupported,
+            class Buffers>
         typename std::enable_if<
             ! http::detail::is_header<Buffers>::value>::type
-        accept(stream<NextLayer>& ws,
+        accept(stream<NextLayer, deflateSupported>& ws,
             Buffers const& buffers) const
         {
             ws.accept(buffers);
         }
 
-        template<class NextLayer>
+        template<class NextLayer, bool deflateSupported>
         void
-        accept(stream<NextLayer>& ws,
+        accept(
+            stream<NextLayer, deflateSupported>& ws,
             http::request<http::empty_body> const& req) const
         {
             ws.accept(req);
         }
 
-        template<class NextLayer, class Decorator>
+        template<
+            class NextLayer, bool deflateSupported,
+            class Decorator>
         void
-        accept_ex(stream<NextLayer>& ws,
+        accept_ex(
+            stream<NextLayer, deflateSupported>& ws,
             Decorator const& d) const
         {
             ws.accept_ex(d);
         }
 
-        template<class NextLayer,
+        template<
+            class NextLayer, bool deflateSupported,
             class Buffers, class Decorator>
         typename std::enable_if<
             ! http::detail::is_header<Buffers>::value>::type
-        accept_ex(stream<NextLayer>& ws,
+        accept_ex(
+            stream<NextLayer, deflateSupported>& ws,
             Buffers const& buffers,
-                Decorator const& d) const
+            Decorator const& d) const
         {
             ws.accept_ex(buffers, d);
         }
 
-        template<class NextLayer, class Decorator>
+        template<
+            class NextLayer, bool deflateSupported,
+            class Decorator>
         void
-        accept_ex(stream<NextLayer>& ws,
+        accept_ex(
+            stream<NextLayer, deflateSupported>& ws,
             http::request<http::empty_body> const& req,
-                Decorator const& d) const
+            Decorator const& d) const
         {
             ws.accept_ex(req, d);
         }
 
-        template<class NextLayer,
+        template<
+            class NextLayer, bool deflateSupported,
             class Buffers, class Decorator>
         void
-        accept_ex(stream<NextLayer>& ws,
+        accept_ex(
+            stream<NextLayer, deflateSupported>& ws,
             http::request<http::empty_body> const& req,
-                Buffers const& buffers,
-                    Decorator const& d) const
+            Buffers const& buffers,
+            Decorator const& d) const
         {
             ws.accept_ex(req, buffers, d);
         }
 
-        template<class NextLayer>
+        template<class NextLayer, bool deflateSupported>
         void
-        handshake(stream<NextLayer>& ws,
+        handshake(
+            stream<NextLayer, deflateSupported>& ws,
             string_view uri,
-                string_view path) const
+            string_view path) const
         {
             ws.handshake(uri, path);
         }
 
-        template<class NextLayer>
+        template<class NextLayer, bool deflateSupported>
         void
-        handshake(stream<NextLayer>& ws,
+        handshake(
+            stream<NextLayer, deflateSupported>& ws,
             response_type& res,
-                string_view uri,
-                    string_view path) const
+            string_view uri,
+            string_view path) const
         {
             ws.handshake(res, uri, path);
         }
 
-        template<class NextLayer, class Decorator>
+        template<
+            class NextLayer, bool deflateSupported,
+            class Decorator>
         void
-        handshake_ex(stream<NextLayer>& ws,
+        handshake_ex(
+            stream<NextLayer, deflateSupported>& ws,
             string_view uri,
-                string_view path,
-                    Decorator const& d) const
+            string_view path,
+            Decorator const& d) const
         {
             ws.handshake_ex(uri, path, d);
         }
 
-        template<class NextLayer, class Decorator>
+        template<
+            class NextLayer, bool deflateSupported,
+            class Decorator>
         void
-        handshake_ex(stream<NextLayer>& ws,
+        handshake_ex(
+            stream<NextLayer, deflateSupported>& ws,
             response_type& res,
-                string_view uri,
-                    string_view path,
-                        Decorator const& d) const
+            string_view uri,
+            string_view path,
+            Decorator const& d) const
         {
             ws.handshake_ex(res, uri, path, d);
         }
 
-        template<class NextLayer>
+        template<class NextLayer, bool deflateSupported>
         void
-        ping(stream<NextLayer>& ws,
+        ping(stream<NextLayer, deflateSupported>& ws,
             ping_data const& payload) const
         {
             ws.ping(payload);
         }
 
-        template<class NextLayer>
+        template<class NextLayer, bool deflateSupported>
         void
-        pong(stream<NextLayer>& ws,
+        pong(stream<NextLayer, deflateSupported>& ws,
             ping_data const& payload) const
         {
             ws.pong(payload);
         }
 
-        template<class NextLayer>
+        template<class NextLayer, bool deflateSupported>
         void
-        close(stream<NextLayer>& ws,
+        close(stream<NextLayer, deflateSupported>& ws,
             close_reason const& cr) const
         {
             ws.close(cr);
         }
 
         template<
-            class NextLayer, class DynamicBuffer>
+            class NextLayer, bool deflateSupported,
+            class DynamicBuffer>
         std::size_t
-        read(stream<NextLayer>& ws,
+        read(stream<NextLayer, deflateSupported>& ws,
             DynamicBuffer& buffer) const
         {
             return ws.read(buffer);
         }
 
         template<
-            class NextLayer, class DynamicBuffer>
+            class NextLayer, bool deflateSupported,
+            class DynamicBuffer>
         std::size_t
-        read_some(stream<NextLayer>& ws,
+        read_some(
+            stream<NextLayer, deflateSupported>& ws,
             std::size_t limit,
             DynamicBuffer& buffer) const
         {
@@ -628,36 +657,45 @@ public:
         }
 
         template<
-            class NextLayer, class MutableBufferSequence>
+            class NextLayer, bool deflateSupported,
+            class MutableBufferSequence>
         std::size_t
-        read_some(stream<NextLayer>& ws,
+        read_some(
+            stream<NextLayer, deflateSupported>& ws,
             MutableBufferSequence const& buffers) const
         {
             return ws.read_some(buffers);
         }
 
         template<
-            class NextLayer, class ConstBufferSequence>
+            class NextLayer, bool deflateSupported,
+            class ConstBufferSequence>
         std::size_t
-        write(stream<NextLayer>& ws,
+        write(
+            stream<NextLayer, deflateSupported>& ws,
             ConstBufferSequence const& buffers) const
         {
             return ws.write(buffers);
         }
 
         template<
-            class NextLayer, class ConstBufferSequence>
+            class NextLayer, bool deflateSupported,
+            class ConstBufferSequence>
         std::size_t
-        write_some(stream<NextLayer>& ws, bool fin,
+        write_some(
+            stream<NextLayer, deflateSupported>& ws,
+            bool fin,
             ConstBufferSequence const& buffers) const
         {
             return ws.write_some(fin, buffers);
         }
 
         template<
-            class NextLayer, class ConstBufferSequence>
+            class NextLayer, bool deflateSupported,
+            class ConstBufferSequence>
         std::size_t
-        write_raw(stream<NextLayer>& ws,
+        write_raw(
+            stream<NextLayer, deflateSupported>& ws,
             ConstBufferSequence const& buffers) const
         {
             return boost::asio::write(
@@ -678,9 +716,9 @@ public:
         {
         }
 
-        template<class NextLayer>
+        template<class NextLayer, bool deflateSupported>
         void
-        accept(stream<NextLayer>& ws) const
+        accept(stream<NextLayer, deflateSupported>& ws) const
         {
             error_code ec;
             ws.async_accept(yield_[ec]);
@@ -688,10 +726,13 @@ public:
                 throw system_error{ec};
         }
 
-        template<class NextLayer, class Buffers>
+        template<
+            class NextLayer, bool deflateSupported,
+            class Buffers>
         typename std::enable_if<
             ! http::detail::is_header<Buffers>::value>::type
-        accept(stream<NextLayer>& ws,
+        accept(
+            stream<NextLayer, deflateSupported>& ws,
             Buffers const& buffers) const
         {
             error_code ec;
@@ -700,9 +741,10 @@ public:
                 throw system_error{ec};
         }
 
-        template<class NextLayer>
+        template<class NextLayer, bool deflateSupported>
         void
-        accept(stream<NextLayer>& ws,
+        accept(
+            stream<NextLayer, deflateSupported>& ws,
             http::request<http::empty_body> const& req) const
         {
             error_code ec;
@@ -711,10 +753,12 @@ public:
                 throw system_error{ec};
         }
 
-        template<class NextLayer,
+        template<
+            class NextLayer, bool deflateSupported,
             class Decorator>
         void
-        accept_ex(stream<NextLayer>& ws,
+        accept_ex(
+            stream<NextLayer, deflateSupported>& ws,
             Decorator const& d) const
         {
             error_code ec;
@@ -723,13 +767,15 @@ public:
                 throw system_error{ec};
         }
 
-        template<class NextLayer,
+        template<
+            class NextLayer, bool deflateSupported,
             class Buffers, class Decorator>
         typename std::enable_if<
             ! http::detail::is_header<Buffers>::value>::type
-        accept_ex(stream<NextLayer>& ws,
+        accept_ex(
+            stream<NextLayer, deflateSupported>& ws,
             Buffers const& buffers,
-                Decorator const& d) const
+            Decorator const& d) const
         {
             error_code ec;
             ws.async_accept_ex(buffers, d, yield_[ec]);
@@ -737,11 +783,14 @@ public:
                 throw system_error{ec};
         }
 
-        template<class NextLayer, class Decorator>
+        template<
+            class NextLayer, bool deflateSupported,
+            class Decorator>
         void
-        accept_ex(stream<NextLayer>& ws,
+        accept_ex(
+            stream<NextLayer, deflateSupported>& ws,
             http::request<http::empty_body> const& req,
-                Decorator const& d) const
+            Decorator const& d) const
         {
             error_code ec;
             ws.async_accept_ex(req, d, yield_[ec]);
@@ -749,13 +798,15 @@ public:
                 throw system_error{ec};
         }
 
-        template<class NextLayer,
+        template<
+            class NextLayer, bool deflateSupported,
             class Buffers, class Decorator>
         void
-        accept_ex(stream<NextLayer>& ws,
+        accept_ex(
+            stream<NextLayer, deflateSupported>& ws,
             http::request<http::empty_body> const& req,
-                Buffers const& buffers,
-                    Decorator const& d) const
+            Buffers const& buffers,
+            Decorator const& d) const
         {
             error_code ec;
             ws.async_accept_ex(
@@ -764,11 +815,13 @@ public:
                 throw system_error{ec};
         }
 
-        template<class NextLayer>
+        template<
+            class NextLayer, bool deflateSupported>
         void
-        handshake(stream<NextLayer>& ws,
+        handshake(
+            stream<NextLayer, deflateSupported>& ws,
             string_view uri,
-                string_view path) const
+            string_view path) const
         {
             error_code ec;
             ws.async_handshake(
@@ -777,12 +830,13 @@ public:
                 throw system_error{ec};
         }
 
-        template<class NextLayer>
+        template<class NextLayer, bool deflateSupported>
         void
-        handshake(stream<NextLayer>& ws,
+        handshake(
+            stream<NextLayer, deflateSupported>& ws,
             response_type& res,
-                string_view uri,
-                    string_view path) const
+            string_view uri,
+            string_view path) const
         {
             error_code ec;
             ws.async_handshake(
@@ -791,12 +845,15 @@ public:
                 throw system_error{ec};
         }
 
-        template<class NextLayer, class Decorator>
+        template<
+            class NextLayer, bool deflateSupported,
+            class Decorator>
         void
-        handshake_ex(stream<NextLayer>& ws,
+        handshake_ex(
+            stream<NextLayer, deflateSupported>& ws,
             string_view uri,
-                string_view path,
-                    Decorator const &d) const
+            string_view path,
+            Decorator const &d) const
         {
             error_code ec;
             ws.async_handshake_ex(
@@ -805,13 +862,16 @@ public:
                 throw system_error{ec};
         }
 
-        template<class NextLayer, class Decorator>
+        template<
+            class NextLayer, bool deflateSupported,
+            class Decorator>
         void
-        handshake_ex(stream<NextLayer>& ws,
+        handshake_ex(
+            stream<NextLayer, deflateSupported>& ws,
             response_type& res,
-                string_view uri,
-                    string_view path,
-                        Decorator const &d) const
+            string_view uri,
+            string_view path,
+            Decorator const &d) const
         {
             error_code ec;
             ws.async_handshake_ex(
@@ -820,9 +880,10 @@ public:
                 throw system_error{ec};
         }
 
-        template<class NextLayer>
+        template<class NextLayer, bool deflateSupported>
         void
-        ping(stream<NextLayer>& ws,
+        ping(
+            stream<NextLayer, deflateSupported>& ws,
             ping_data const& payload) const
         {
             error_code ec;
@@ -831,9 +892,10 @@ public:
                 throw system_error{ec};
         }
 
-        template<class NextLayer>
+        template<class NextLayer, bool deflateSupported>
         void
-        pong(stream<NextLayer>& ws,
+        pong(
+            stream<NextLayer, deflateSupported>& ws,
             ping_data const& payload) const
         {
             error_code ec;
@@ -842,9 +904,10 @@ public:
                 throw system_error{ec};
         }
 
-        template<class NextLayer>
+        template<class NextLayer, bool deflateSupported>
         void
-        close(stream<NextLayer>& ws,
+        close(
+            stream<NextLayer, deflateSupported>& ws,
             close_reason const& cr) const
         {
             error_code ec;
@@ -854,9 +917,11 @@ public:
         }
 
         template<
-            class NextLayer, class DynamicBuffer>
+            class NextLayer, bool deflateSupported,
+            class DynamicBuffer>
         std::size_t
-        read(stream<NextLayer>& ws,
+        read(
+            stream<NextLayer, deflateSupported>& ws,
             DynamicBuffer& buffer) const
         {
             error_code ec;
@@ -868,9 +933,11 @@ public:
         }
 
         template<
-            class NextLayer, class DynamicBuffer>
+            class NextLayer, bool deflateSupported,
+            class DynamicBuffer>
         std::size_t
-        read_some(stream<NextLayer>& ws,
+        read_some(
+            stream<NextLayer, deflateSupported>& ws,
             std::size_t limit,
             DynamicBuffer& buffer) const
         {
@@ -883,9 +950,11 @@ public:
         }
 
         template<
-            class NextLayer, class MutableBufferSequence>
+            class NextLayer, bool deflateSupported,
+            class MutableBufferSequence>
         std::size_t
-        read_some(stream<NextLayer>& ws,
+        read_some(
+            stream<NextLayer, deflateSupported>& ws,
             MutableBufferSequence const& buffers) const
         {
             error_code ec;
@@ -897,9 +966,11 @@ public:
         }
 
         template<
-            class NextLayer, class ConstBufferSequence>
+            class NextLayer, bool deflateSupported,
+            class ConstBufferSequence>
         std::size_t
-        write(stream<NextLayer>& ws,
+        write(
+            stream<NextLayer, deflateSupported>& ws,
             ConstBufferSequence const& buffers) const
         {
             error_code ec;
@@ -911,9 +982,12 @@ public:
         }
 
         template<
-            class NextLayer, class ConstBufferSequence>
+            class NextLayer, bool deflateSupported,
+            class ConstBufferSequence>
         std::size_t
-        write_some(stream<NextLayer>& ws, bool fin,
+        write_some(
+            stream<NextLayer, deflateSupported>& ws,
+            bool fin,
             ConstBufferSequence const& buffers) const
         {
             error_code ec;
@@ -925,9 +999,11 @@ public:
         }
 
         template<
-            class NextLayer, class ConstBufferSequence>
+            class NextLayer, bool deflateSupported,
+            class ConstBufferSequence>
         std::size_t
-        write_raw(stream<NextLayer>& ws,
+        write_raw(
+            stream<NextLayer, deflateSupported>& ws,
             ConstBufferSequence const& buffers) const
         {
             error_code ec;